requirements.txt
flask
mysql.connector

